--- ascii_art/__init__.py
@@ -1,3 +1,0 @@
-import ascii_art.chem
-import ascii_art.title
-import ascii_art.thank_you
