+++ src/ascii_art/__init__.py
@@ -1,0 +1,3 @@
+import src.ascii_art.chem
+import src.ascii_art.title
+import src.ascii_art.thank_you
