+++ src/ascii_art/thank_you.py
