+++ src/configs.py
@@ -1,8 +1,6 @@
 import cript as C
 import inspect
-from util import filter_required_col
-
-BASE_URL = "https://criptapp-staging.herokuapp.com/"
+from src.util import filter_required_col
 
 required_cols = {
     "experiment": filter_required_col(C.Experiment.required),
